+++ app/controllers/mokio/backend_search_controller.rb
@@ -1,0 +1,3 @@
+class Mokio::BackendSearchController < Mokio::CommonController
+    include Mokio::Concerns::Controllers::BackendSearch
+end