+++ app/helpers/mokio/backend/common_helper.rb
@@ -14,8 +14,9 @@ module Mokio
         obj.respond_to?("data_file_type") ? obj.data_file_type.to_s.demodulize.tableize : "data_files"
       end
 
-      def responsive_dynamic_table(&block)
-        content_tag :table, :data => {:source => datatable_source_url}, :class => "responsive dynamicTable display table table-bordered", :border => 0, :cellpadding => 0, :cellspacing => 0, :width => "100%", :id =>"dTable" do
+      def responsive_dynamic_table(without_search = false,&block)
+        without_search_class = (without_search) ? 'without-search': ''
+        content_tag :table, :data => {:source => datatable_source_url}, :class => "responsive dynamicTable display table table-bordered #{without_search_class}", :border => 0, :cellpadding => 0, :cellspacing => 0, :width => "100%", :id =>"dTable" do
           capture(&block)
         end
       end