+++ app/models/mokio/backend_search.rb
@@ -1,0 +1,5 @@
+module Mokio
+    class BackendSearch
+        include Mokio::Concerns::Models::BackendSearch
+    end
+end